+++ crates/tx-graph/src/peg_out_graph.rs
@@ -642,7 +642,7 @@ mod tests {
         Transaction, TxOut,
     };
     use bitcoind_async_client::types::GetTxOut;
-    use bitvm::signatures::wots_api::HASH_LEN;
+    use bitvm::signatures::HASH_LEN;
     use corepc_node::{serde_json::json, Client, Conf, Node};
     use rkyv::rancor::Error;
     use secp256k1::{
@@ -1049,8 +1049,8 @@ mod tests {
         for _ in 0..faulty_assertions.groth16.2.len() {
             let proof_index_to_tweak = OsRng.gen_range(0..faulty_assertions.groth16.2.len());
             warn!(action = "introducing faulty assertion", index=%proof_index_to_tweak);
-            if faulty_assertions.groth16.2[proof_index_to_tweak] != [0u8; HASH_LEN as usize] {
-                faulty_assertions.groth16.2[proof_index_to_tweak] = [0u8; HASH_LEN as usize];
+            if faulty_assertions.groth16.2[proof_index_to_tweak] != [0u8; HASH_LEN] {
+                faulty_assertions.groth16.2[proof_index_to_tweak] = [0u8; HASH_LEN];
                 break;
             }
         }
